Move more functions to subpackages
